whisper/requirements.txt
openai-whisper==20231117
minio==7.2.0
python-dotenv==1.0.0
torch==2.1.0
torchaudio==2.1.0
ffmpeg-python==0.2.0

